requirements.txt
Django==4.0.2
Pillow==8.4.0
django-environ==0.8.1
requests==2.27.1
gunicorn==20.1.0
celery==5.1.2
redis==3.5.3
pyjwt==2.3.0
python-telegram-bot==13.8.1
sentry-sdk==1.5.0
tweepy==4.5.0
django-simple-history==3.0.0
django-json-widget==1.1.1
boto3==1.21.18
botocore==1.24.18
nanoid==2.0.0

# patreon bot
quart==0.16.2
patreon==0.5.0
aiogram==2.18
apscheduler==3.6.3
pytz==2021.3
MarkupSafe==2.0.1
hypercorn==0.13.2
telegraph==2.1.0